Some small changes.

+++ src/Utils/Constants.hpp
@@ -5,8 +5,8 @@ namespace Game
 	namespace Constants
 	{
 		constexpr char TITLE[] = "Aetherwild";
-		constexpr int SCREEN_WIDTH = 800;
-		constexpr int SCREEN_HEIGHT = 600;
-		constexpr int FPS = 60;
+		constexpr uint16_t SCREEN_WIDTH = 800;
+		constexpr uint16_t SCREEN_HEIGHT = 600;
+		constexpr uint8_t FPS = 60;
 	}
 }
